@@ -11,38 +11,37 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.15",
-    "@angular/common": "^7.2.15",
-    "@angular/compiler": "^7.2.15",
-    "@angular/core": "^7.2.15",
-    "@angular/forms": "^7.2.15",
-    "@angular/http": "^7.2.15",
-    "@angular/platform-browser": "^7.2.15",
-    "@angular/platform-browser-dynamic": "^7.2.15",
-    "@angular/router": "^7.2.15",
+    "@angular/animations": "^8.2.14",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
     "@fortawesome/fontawesome-free": "^5.10.0",
-    "@ng-select/ng-select": "^2.5.0",
-    "@trademe/ng-defer-load": "^3.0.1",
+    "@ng-select/ng-select": "^3.7.0",
     "@types/chart.js": "^2.7.56",
-    "angular-bootstrap-md": "^7.5.4",
+    "angular-bootstrap-md": "^8.8.1",
     "animate.css": "^3.7.2",
-    "chart.js": "^2.5.0",
-    "core-js": "^2.5.4",
+    "chart.js": "^2.9.3",
+    "core-js": "^3.6.1",
     "font-awesome": "^4.7.0",
     "hammerjs": "^2.0.8",
+    "ng-lazyload-image": "^7.0.1",
     "node-sass": "^4.13.0",
     "rxjs": "^6.3.3",
-    "zone.js": "^0.8.26"
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.13.9",
-    "@angular/cli": "^7.3.9",
-    "@angular/compiler-cli": "^7.2.15",
-    "@angular/language-service": "^6.1.3",
-    "@types/jasmine": "^2.8.8",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^8.9.5",
-    "codelyzer": "5.1.0",
+    "@angular-devkit/build-angular": "0.803.21",
+    "@angular/cli": "^8.3.20",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "^8.2.14",
+    "@types/jasmine": "^3.5.0",
+    "@types/jasminewd2": "~2.0.8",
+    "@types/node": "^13.1.2",
+    "codelyzer": "5.2.1",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "^4.2.0",
@@ -53,7 +52,7 @@
     "protractor": "^5.4.1",
     "ts-node": "~5.0.1",
     "tslib": "^1.9.3",
-    "tslint": "5.18.0",
-    "typescript": "3.2.4"
+    "tslint": "5.20.1",
+    "typescript": "3.5.3"
   }
 }
