@@ -11,11 +11,12 @@ import { DownloadDialogComponent } from '../download-dialog/download-dialog.comp
 
 @Injectable()
 export class PackageInfoComponent {
+  public defaultImage = 'assets/package.png';
 
   @Input()
   public package: PackageDTO;
 
-  @ViewChild(DownloadDialogComponent) downloadDialog: DownloadDialogComponent;
+  @ViewChild(DownloadDialogComponent, { static: true }) downloadDialog: DownloadDialogComponent;
   showDownloadDialogModal() {
     this.downloadDialog.showModal();
   }
