@@ -6,7 +6,7 @@ namespace synopackage_dotnet.Model.Services
 {
   public interface IDownloadService
   {
-    Task<ExecuteResponse> Execute(string url, IEnumerable<KeyValuePair<string, object>> parameters, string userAgent = null);
+    Task<ExecuteResponse> Execute(string url, IEnumerable<KeyValuePair<string, object>> parameters, string userAgent = null, bool useGetMethod = false);
     Task<byte[]> DownloadData(string url);
   }
 }
